Update LAB_01L01_Prepare_for_deployment_of_AVD_ADDS.md
Exercise 1
Task 2 Step 3 Domain Controller deployment from template

We should modify the template which deploys the Domain Controller with the following:
Remove the current password as it is on the attackers' list. 
Let each class participant enter a unique password for the "Student " account. 
In my recent delivery, about 7 students complained that there password got changed.